+++ ios-device-lib/Classes/Terminals/UPA/Builders/HpsUpaAuthBuilder.m
@@ -42,6 +42,10 @@
         request.data.data.transaction.referenceNumber = self.transactionId;
     }
     
+    if (self.allowDuplicate != nil) {
+        request.data.data.transaction.allowDuplicate = self.allowDuplicate;
+    }
+    
     request.data.data.transaction.clinicAmount = self.clinicAmount != nil ? [formatter stringFromNumber:[NSNumber numberWithDouble:[self.clinicAmount doubleValue]]] : nil;
     request.data.data.transaction.prescriptionAmount = self.prescriptionAmount != nil ? [formatter stringFromNumber:[NSNumber numberWithDouble:[self.prescriptionAmount doubleValue]]] : nil;
     request.data.data.transaction.dentalAmount = self.dentalAmount != nil ? [formatter stringFromNumber:[NSNumber numberWithDouble:[self.dentalAmount doubleValue]]] : nil;