+++ ios-device-lib/Classes/Terminals/UPA/Constants/Upa_Constants.h
@@ -1,0 +1,11 @@
+//
+//  Upa_Constants.h
+//  ios-device-lib
+//
+
+#import <Foundation/Foundation.h>
+
+@interface Upa_Constants : NSObject
+extern NSString * const username;
+extern NSString * const password;
+@end