+++ ios-device-lib/Classes/Terminals/UPA/Constants/Upa_Constants.m
@@ -1,0 +1,11 @@
+//
+//  Upa_Constants.m
+//  ios-device-lib
+//
+
+#import "Upa_Constants.h"
+
+@implementation Upa_Constants : NSObject
+NSString * const username = @"701420636";
+NSString * const password = @"$Test1234";
+@end